LPS-142497 Move Topper components to their own folder
