+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/app/components/layout-data-items/CollectionItemWithControls.js
@@ -16,7 +16,7 @@ import classNames from 'classnames';
 import React, {useContext} from 'react';
 
 import {CollectionItemContext} from '../../contexts/CollectionItemContext';
-import TopperEmpty from '../TopperEmpty';
+import TopperEmpty from '../topper/TopperEmpty';
 
 const CollectionItemWithControls = React.forwardRef(({children, item}, ref) => {
 	const {collectionItem} = useContext(CollectionItemContext);
