+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/app/components/layout-data-items/ColumnWithControls.js
@@ -38,7 +38,7 @@ import {
 import {getResponsiveColumnSize} from '../../utils/getResponsiveColumnSize';
 import {getResponsiveConfig} from '../../utils/getResponsiveConfig';
 import isItemEmpty from '../../utils/isItemEmpty';
-import TopperEmpty from '../TopperEmpty';
+import TopperEmpty from '../topper/TopperEmpty';
 import Column from './Column';
 
 /**
