+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/app/components/layout-data-items/ContainerWithControls.js
@@ -27,7 +27,7 @@ import selectCanUpdateItemConfiguration from '../../selectors/selectCanUpdateIte
 import {getFrontendTokenValue} from '../../utils/getFrontendTokenValue';
 import {getResponsiveConfig} from '../../utils/getResponsiveConfig';
 import {isValidSpacingOption} from '../../utils/isValidSpacingOption';
-import Topper from '../Topper';
+import Topper from '../topper/Topper';
 import Container from './Container';
 import isHovered from './isHovered';
