+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/app/components/layout-data-items/RowWithControls.js
@@ -25,7 +25,7 @@ import {getResponsiveColumnSize} from '../../utils/getResponsiveColumnSize';
 import {getResponsiveConfig} from '../../utils/getResponsiveConfig';
 import isItemEmpty from '../../utils/isItemEmpty';
 import {isValidSpacingOption} from '../../utils/isValidSpacingOption';
-import Topper from '../Topper';
+import Topper from '../topper/Topper';
 import Row from './Row';
 
 const ROW_SIZE = 12;
