+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/app/components/topper/TopperEmpty.js
@@ -15,11 +15,11 @@
 import classNames from 'classnames';
 import React, {useRef} from 'react';
 
-import {getLayoutDataItemPropTypes} from '../../prop-types/index';
-import {useSelector} from '../contexts/StoreContext';
-import selectCanUpdatePageStructure from '../selectors/selectCanUpdatePageStructure';
-import {TARGET_POSITIONS} from '../utils/drag-and-drop/constants/targetPositions';
-import {useDropTarget} from '../utils/drag-and-drop/useDragAndDrop';
+import {getLayoutDataItemPropTypes} from '../../../prop-types/index';
+import {useSelector} from '../../contexts/StoreContext';
+import selectCanUpdatePageStructure from '../../selectors/selectCanUpdatePageStructure';
+import {TARGET_POSITIONS} from '../../utils/drag-and-drop/constants/targetPositions';
+import {useDropTarget} from '../../utils/drag-and-drop/useDragAndDrop';
 
 export default function ({children, ...props}) {
 	const canUpdatePageStructure = useSelector(selectCanUpdatePageStructure);
