+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/app/components/topper/_Topper.scss
@@ -1,4 +1,4 @@
-@import './variables';
+@import '../variables';
 
 $borderWidth: 2px;
 $color: $white;
